backend/pyproject.toml
[project]
name = "whisperworld-backend"
version = "0.1.0"
description = "WhisperWorld FastAPI backend — voice conversations with nature"
requires-python = ">=3.12"
dependencies = [
    "fastapi",
    "uvicorn[standard]",
    "python-dotenv",
    "supabase",
    "httpx",
    "websockets",
    "google-generativeai",
    "pydantic",
]

[project.optional-dependencies]
dev = [
    "pytest",
    "pytest-asyncio",
    "httpx",
]

[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.backends.legacy:build"

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
